requirements.txt
google-adk>=0.5.0
python-dotenv>=1.0.0
litellm>=1.50.0
